add report

+++ A2_Evolutionary-Algorithms/Instance.java
@@ -2,12 +2,12 @@ class Instance {
     int numItems;
     int capacity;
 
-    int[] values;
-    int[] weights;
+    double[] values;
+    double[] weights;
 
     String solution;
 
-    Instance(int numItems, int capacity, int[] values, int[] weights, String solution) {
+    Instance(int numItems, int capacity, double[] values, double[] weights, String solution) {
         this.values = values;
         this.weights = weights;
         this.solution = solution;
